+++ app/src/main/java/co/revely/peertube/api/peertube/response/File.kt
@@ -3,8 +3,8 @@ package co.revely.peertube.api.peertube.response
 data class File(
 		val resolution: Label?,
 		val magnetUri: String?,
-		val size: Long?,
-		val fps: Long?,
+		val size: Int?,
+		val fps: Int?,
 		val torrentUrl: String?,
 		val torrentDownloadUrl: String?,
 		val fileUrl: String?,
