+++ app/src/main/java/co/revely/peertube/api/peertube/response/User.kt
@@ -6,15 +6,15 @@ package co.revely.peertube.api.peertube.response
  * @author rbenjami
  */
 data class User(
-		val id: Long?,
+		val id: Int?,
 		val username: String?,
 		val email: String?,
 		val displayNSFW: Boolean?,
 		val autoPlayVideo: Boolean?,
-		val role: Long?,
+		val role: Int?,
 		val roleLabel: String?,
-		val videoQuota: Long?,
-		val videoQuotaDaily: Long?,
+		val videoQuota: Int?,
+		val videoQuotaDaily: Int?,
 		val createdAt: String?,
 		var account: Account?,
 		var videoChannels: List<Channel>?
